+++ controllers/entry.controller.ts
@@ -3,7 +3,7 @@ import { connectDB } from "../lib/mongodb"
 import { Collection, DeleteResult, InsertOneResult, MongoClient, ObjectId, UpdateResult } from "mongodb"
 
 //Interface
-import { Entry } from "../interfaces"
+import { Entry } from "../interfaces/entry"
 
 export class EntryController {
   entry: Entry
@@ -14,7 +14,7 @@ export class EntryController {
   }
 
   async create() {
-    let client: MongoClient
+    let client: MongoClient | undefined
     let dbCollection: Collection
     let isConnected = false
 
@@ -24,7 +24,7 @@ export class EntryController {
     } catch (e) {
       console.log(e)
     }
-    if (isConnected) {
+    if (client && isConnected) {
       let insertResult: InsertOneResult
       try {
         dbCollection = client.db(process.env.DB_NAME).collection("entries")
@@ -55,7 +55,7 @@ export class EntryController {
   }
 
   async update(id: string) {
-    let client: MongoClient
+    let client: MongoClient | undefined
     let dbCollection: Collection
     let isConnected = false
 
@@ -65,8 +65,8 @@ export class EntryController {
     } catch (e) {
       console.log(e)
     }
-    if (isConnected) {
-      let updateResult: UpdateResult
+    if (client && isConnected) {
+      let updateResult: UpdateResult | undefined
       try {
         dbCollection = client.db(process.env.DB_NAME).collection("entries")
         updateResult = await dbCollection.updateOne({ _id: new ObjectId(id) }, { $set: this.entry }, { upsert: false })
@@ -77,9 +77,9 @@ export class EntryController {
           await client.close()
         }*/
       }
-      if (updateResult.modifiedCount === 1) {
+      if (updateResult && updateResult.modifiedCount === 1) {
         return { status: "success", message: "Entry has been updated." }
-      } else if (updateResult.matchedCount === 1 && updateResult.modifiedCount === 0) {
+      } else if (updateResult && updateResult.matchedCount === 1 && updateResult.modifiedCount === 0) {
         return { status: "failed", message: "You didn't make any change." }
       } else {
         return { status: "failed", message: "Failed to update the entry." }
@@ -90,7 +90,7 @@ export class EntryController {
   }
 
   async delete(id: string) {
-    let client: MongoClient
+    let client: MongoClient | undefined
     let dbCollection: Collection
     let isConnected: boolean = false
 
@@ -100,8 +100,8 @@ export class EntryController {
     } catch (e) {
       console.log(e)
     }
-    if (isConnected) {
-      let deleteResult: DeleteResult
+    if (client && isConnected) {
+      let deleteResult: DeleteResult | undefined
       try {
         dbCollection = client.db(process.env.DB_NAME).collection("entries")
         deleteResult = await dbCollection.deleteOne({ _id: new ObjectId(id) })
@@ -112,7 +112,7 @@ export class EntryController {
           await client.close()
         }*/
       }
-      if (deleteResult.deletedCount === 1) {
+      if (deleteResult && deleteResult.deletedCount === 1) {
         return { status: "success", message: "Entry has been deleted." }
       } else {
         return { status: "failed", message: "Failed to delete the entry." }
